restructure the directory (#3431)
