+++ src/AutoRest.CSharp/Common/Input/InputTypes/BodyMediaType.cs
@@ -1,0 +1,15 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+namespace AutoRest.CSharp.Common.Input;
+
+internal enum BodyMediaType
+{
+    None,
+    Binary,
+    Form,
+    Json,
+    Multipart,
+    Text,
+    Xml
+}
