+++ src/AutoRest.CSharp/Common/Input/InputTypes/InputApiKeyAuth.cs
@@ -1,0 +1,12 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System;
+using System.Collections.Generic;
+
+namespace AutoRest.CSharp.Common.Input;
+
+internal record InputApiKeyAuth(string Name)
+{
+    public InputApiKeyAuth() : this(string.Empty) { }
+}
