+++ src/AutoRest.CSharp/Common/Input/InputTypes/InputConstant.cs
@@ -1,0 +1,6 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+namespace AutoRest.CSharp.Common.Input;
+
+internal record InputConstant(object? Value, InputType Type);
