+++ src/AutoRest.CSharp/Common/Input/InputTypes/InputIntrinsicTypeKind.cs
@@ -1,0 +1,13 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+namespace AutoRest.CSharp.Common.Input;
+
+internal enum InputIntrinsicTypeKind
+{
+    ErrorType,
+    Void,
+    Never,
+    Unknown,
+    Null,
+}
