+++ src/AutoRest.CSharp/Common/Input/InputTypes/InputParameter.cs
@@ -1,0 +1,51 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using AutoRest.CSharp.Input;
+using AutoRest.CSharp.Output.Models.Serialization;
+
+namespace AutoRest.CSharp.Common.Input;
+
+internal record InputParameter(
+    string Name,
+    string NameInRequest,
+    string? Description,
+    InputType Type,
+    RequestLocation Location,
+    InputConstant? DefaultValue,
+    VirtualParameter? VirtualParameter,
+    InputParameter? GroupedBy,
+    InputOperationParameterKind Kind,
+    bool IsRequired,
+    bool IsApiVersion,
+    bool IsResourceParameter,
+    bool IsContentType,
+    bool IsEndpoint,
+    bool SkipUrlEncoding,
+    bool Explode,
+    string? ArraySerializationDelimiter,
+    string? HeaderCollectionPrefix,
+    SerializationFormat SerializationFormat)
+{
+    public InputParameter() : this(
+        Name: string.Empty,
+        NameInRequest: string.Empty,
+        Description: null,
+        Type: new InputPrimitiveType(InputTypeKind.Object),
+        Location: RequestLocation.None,
+        DefaultValue: null,
+        VirtualParameter: null,
+        GroupedBy: null,
+        Kind: InputOperationParameterKind.Method,
+        IsRequired: false,
+        IsApiVersion: false,
+        IsResourceParameter: false,
+        IsContentType: false,
+        IsEndpoint: false,
+        SkipUrlEncoding: false,
+        Explode: false,
+        ArraySerializationDelimiter: null,
+        HeaderCollectionPrefix: null,
+        SerializationFormat: SerializationFormat.Default)
+    { }
+}
