+++ src/AutoRest.CSharp/Common/Input/InputTypes/InputType.cs
@@ -1,0 +1,6 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+namespace AutoRest.CSharp.Common.Input;
+
+internal abstract record InputType(string Name, bool IsNullable = false) { }
