+++ src/AutoRest.CSharp/Common/Input/InputTypes/InputTypeKind.cs
@@ -1,0 +1,43 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+namespace AutoRest.CSharp.Common.Input;
+
+internal enum InputTypeKind
+{
+    AzureLocation,
+    Boolean,
+    BinaryData,
+    Bytes,
+    BytesBase64Url,
+    ContentType,
+    Date,
+    DateTime,
+    DateTimeISO8601,
+    DateTimeRFC1123,
+    DateTimeRFC3339,
+    DateTimeRFC7231,
+    DateTimeUnix,
+    DurationISO8601,
+    DurationConstant,
+    DurationSeconds,
+    DurationSecondsFloat,
+    ETag,
+    Float32,
+    Float64,
+    Float128,
+    Guid,
+    Int32,
+    Int64,
+    IPAddress,
+    Object,
+    RequestMethod,
+    ResourceIdentifier,
+    ResourceType,
+    Stream,
+    String,
+    Time,
+    Uri,
+    SByte,
+    Byte
+}
