+++ src/AutoRest.CSharp/Common/Input/InputTypes/RequestLocation.cs
