+++ src/AutoRest.CSharp/Common/Input/Utf8JsonReaderExtensions.cs
@@ -123,7 +123,7 @@ namespace AutoRest.CSharp.Common.Input
             }
 
             reader.Read();
-            value = CadlInputTypeConverter.CreatePrimitiveType(reader.GetString(), false) ?? throw new JsonException();
+            value = TypespecInputTypeConverter.CreatePrimitiveType(reader.GetString(), false) ?? throw new JsonException();
             reader.Read();
             return true;
         }
